add notes on download impelmentation

+++ core/datasets/pfdelta_dataset.py
@@ -1,5 +1,6 @@
 import json
 import os
+import ipdb
 from tqdm import tqdm
 from functools import partial
 import glob
@@ -20,9 +21,9 @@ class PFDeltaDataset(InMemoryDataset):
     def __init__(
         self,
         root_dir="data",
-        case_name="",
-        perturbation="",
-        feasibility_type="",
+        case_name="case14",
+        perturbation="n",
+        feasibility_type="feasible",
         n_samples=-1,
         split="train",
         model="",
@@ -79,8 +80,8 @@ class PFDeltaDataset(InMemoryDataset):
             2.2: {"feasible": {"n": 12000, "n-1": 12000, "n-2": 12000}},
             2.3: {"feasible": {"n": 6000, "n-1": 6000, "n-2": 6000}},
             3.1: {"feasible": {"n": 18000, "n-1": 18000, "n-2": 18000}},
-            3.2: {"feasible":{"none": 6000, "n-1": 6000, "n-2": 6000}},
-            3.3: {"feasible":{"none": 9000, "n-1": 9000, "n-2": 9000}},
+            3.2: {"feasible": {"none": 6000, "n-1": 6000, "n-2": 6000}},
+            3.3: {"feasible": {"none": 9000, "n-1": 9000, "n-2": 9000}},
             4.1: {
                 "near infeasible": {"n": 1800, "n-1": 1800, "n-2": 1800},
                 "feasible": {"n": 16200, "n-1": 16200, "n-2": 16200},
@@ -171,10 +172,25 @@ class PFDeltaDataset(InMemoryDataset):
 
         return ["train.pt", "val.pt", "test.pt"]
 
-    def download(self, url="https://huggingface.co/datasets/pfdelta/pfdelta/tree/main"):
+    def _download(self):
+        files = self.files_missing()
+
+        if files:
+            print(f"Files for task {self.task} already downloaded. Skipping download.")
+            return
+
+        # format URLs to download + extract tar files
+        ipdb.set_trace()
         archive_path = download_url(url, self.root, log=True)
         extract_tar(archive_path, self.raw_dir)
         os.unlink(archive_path)
+
+    def files_missing(self):
+        # check for the specified task if all files exist
+        # download the partial files if only half the files asked -- checking folder names
+        # return a set of folder names to download based on the URL based on whats missing
+        # else return None
+        pass
 
     def build_heterodata(self, pm_case, is_cpf_sample=False):
         data = HeteroData()
@@ -644,12 +660,13 @@ class PFDeltaDataset(InMemoryDataset):
         # determine roots based on task
         if task in [3.1, 3.2, 3.3]:
             roots = [
-                os.path.join(self.root, case_name) for case_name in self.all_case_names[:-1]
+                os.path.join(self.root, case_name)
+                for case_name in self.all_case_names[:-1]
             ]
             casename = self.case_name if task == 3.1 else ""
         else:
             roots = [os.path.join(self.root, self.case_name)]
-            casename = self.case_name   
+            casename = self.case_name
 
         if task == "analysis":  # no need to combine anything here
             print(f"Processing data for task {task}")
